Modify CheckingAccount's amount/toamount from int32 to decimal

+++ hMoney/CheckingAccount.cs
@@ -23,8 +23,8 @@ namespace hMoney
         public String PayeeName;
         public int PayeeId;
         public int TransId;
-        public int TransAmount;
-        public int ToTransAmount;
+        public decimal TransAmount;
+        public decimal ToTransAmount;
         public String TransCode;
 
         public String Notes;
